pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "labelrag"
version = "0.0.0"
description = "A label-driven RAG pipeline built on top of paralabelgen."
readme = "README.md"
license = { text = "MIT" }
authors = [
  { name = "huruilizhen", email = "huruilizhen@gmail.com" },
]
requires-python = ">=3.11"
dependencies = [
  "paralabelgen==0.2.0",
]

[project.optional-dependencies]
dev = [
  "pyright>=1.1.390",
  "pytest>=8.3.0",
  "pytest-cov>=6.0.0",
  "ruff>=0.11.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/labelrag"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = [
  "--strict-config",
  "--strict-markers",
]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pyright]
include = ["src", "tests"]
pythonVersion = "3.11"
typeCheckingMode = "strict"
reportMissingTypeStubs = false
venvPath = "."
venv = ".venv"
